- use long everywhere

+++ src/main/java/frigo/math/ExtendedEuclidean.java
@@ -13,38 +13,38 @@ import com.google.common.collect.Lists;
 
 public class ExtendedEuclidean {
 
-    public static int nFromQCompact (List<Integer> q) {
-        int r0 = 0;
-        int r1 = 1;
+    public static long nFromQCompact (List<Long> q) {
+        long r0 = 0;
+        long r1 = 1;
         for( int i = 0; i < q.size(); i++ ){
-            int r2 = r0 + r1 * q.get(i);
+            long r2 = r0 + r1 * q.get(i);
             r0 = r1;
             r1 = r2;
         }
         return r1;
     }
 
-    public static int kFromQCompact (List<Integer> q) {
-        int r0 = 0;
-        int r1 = 1;
+    public static long kFromQCompact (List<Long> q) {
+        long r0 = 0;
+        long r1 = 1;
         for( int i = 0; i < q.size(); i++ ){
-            int r2 = r0 + r1 * q.get(i);
+            long r2 = r0 + r1 * q.get(i);
             r0 = r1;
             r1 = r2;
         }
         return r0;
     }
 
-    public List<Integer> q;
-    public List<Integer> r;
-    public List<Integer> s;
-    public List<Integer> t;
+    public List<Long> q;
+    public List<Long> r;
+    public List<Long> s;
+    public List<Long> t;
 
-    public ExtendedEuclidean (int a, int b) {
-        q = Lists.newArrayList(0);
+    public ExtendedEuclidean (long a, long b) {
+        q = Lists.newArrayList(0L);
         r = Lists.newArrayList(a, b);
-        s = Lists.newArrayList(1, 0);
-        t = Lists.newArrayList(0, 1);
+        s = Lists.newArrayList(1L, 0L);
+        t = Lists.newArrayList(0L, 1L);
     }
 
     public void run () {
@@ -54,35 +54,35 @@ public class ExtendedEuclidean {
             s.add(s.get(i - 1) - s.get(i) * q.get(i));
             t.add(t.get(i - 1) - t.get(i) * q.get(i));
         }
-        q.add(0);
+        q.add(0L);
     }
 
-    public int a () {
+    public long a () {
         return r.get(0);
     }
 
-    public int b () {
+    public long b () {
         return r.get(1);
     }
 
-    public int gcd () {
+    public long gcd () {
         return r.get(r.size() - 2);
     }
 
-    public int binvSigned () {
+    public long binvSigned () {
         return t.get(t.size() - 2);
     }
 
-    public int binv () {
+    public long binv () {
         return (t.get(t.size() - 2) + a()) % a();
     }
 
-    public int k () {
+    public long k () {
         return s.get(s.size() - 2);
     }
 
-    public List<Integer> qshort () {
-        List<Integer> shortq = new ArrayList<>(q);
+    public List<Long> qshort () {
+        List<Long> shortq = new ArrayList<>(q);
         shortq.remove(0);
         shortq.remove(shortq.size() - 1);
         return shortq;
